requirements.txt
appdirs==1.4.4
asyncmy==0.2.3
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==8.0.3
clintermission==0.2.0
cloudconvert==2.0.0
colorama==0.4.4
cryptography==3.2.1
Deprecated==1.2.13
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.0.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Misaka==1.0.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
ftputil==5.0.2
git-delete-merged-branches==6.0.5
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.10.0
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.3
keyring==23.5.0
lxml==4.6.1
Markdown==3.3.6
MarkupSafe==2.0.1
misaka==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
paramiko==2.4.1
prompt-toolkit==3.0.18
protobuf==3.19.4
psutil==5.8.0
pyasn1==0.4.8
pycparser==2.20
pyftpsync==3.1.0
PyGithub==1.55
PyJWT==2.3.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pyodbc==4.0.32
pyparsing==2.4.7
pyperclip==1.8.1
PyQt5==5.15.1
PyQt5-sip==12.8.1
pysftp==0.2.9
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2020.4
PyYAML==6.0
requests==2.25.0
sftpclone==1.2.2
six==1.15.0
smmap==5.0.0
SQLAlchemy==1.4.27
SQLAlchemy-Utils==0.37.9
typing_extensions==4.0.1
urllib3==1.26.2
waitress==2.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
wtforms-html5==0.2.2
xlrd==1.2.0
zipp==3.7.0
